dashboard/requirements.txt
streamlit
pandas
pydeck
psycopg2-binary
requests
python-dotenv
